[viz] Rename phi.app to phi.viz
