+++ phi/flow.py
@@ -3,7 +3,7 @@
 *Main PhiFlow import:* `from phi.flow import *`
 
 Imports important functions and classes from
-`math`, `geom`, `field`, `physics` and `app` (including sub-modules)
+`math`, `geom`, `field`, `physics` and `viz` (including sub-modules)
 as well as the modules and sub-modules themselves.
 
 See `phi.tf.flow`, `phi.torch.flow`, `phi.jax.flow`.
@@ -13,7 +13,7 @@ See `phi.tf.flow`, `phi.torch.flow`, `phi.jax.flow`.
 import numpy
 import numpy as np
 import phi
-from . import math, geom, field, physics, app
+from . import math, geom, field, physics, viz
 from .math import extrapolation, backend
 from .physics import fluid, flip, advect, diffuse
 
@@ -22,7 +22,7 @@ from .math import DType
 from .geom import Geometry, Sphere, Box
 from .field import Grid, CenteredGrid, StaggeredGrid, GeometryMask, SoftGeometryMask, HardGeometryMask, Noise, PointCloud, Scene
 from .physics import Domain, Obstacle
-from .app import App, EditableInt, EditableBool, EditableFloat, EditableString, view
+from .viz import App, EditableInt, EditableBool, EditableFloat, EditableString, view
 
 # Constants
 from .math import PI
@@ -31,7 +31,7 @@ from .physics import OPEN, CLOSED, PERIODIC
 # Functions
 from .math import wrap, tensor, shape
 from .geom import union
-from .app import show
+from .viz import show
 
 # Exceptions
 from .math import ConvergenceException, NotConverged, Diverged
