+++ phi/viz/_console/_console_plot.py
