+++ phi/viz/_console/_console_util.py
