+++ phi/viz/_dash/viewsettings.py
@@ -4,7 +4,7 @@ import dash_core_components as dcc
 import dash_html_components as html
 from dash.dependencies import Input, Output
 
-from phi.app import App
+from phi.viz import App
 from phi.field import Field, CenteredGrid, StaggeredGrid
