+++ phi/viz/_matplotlib/matplotlib_gui.py
