+++ setup.py
@@ -131,8 +131,8 @@ setup(
     version=version,
     download_url='https://github.com/tum-pbs/PhiFlow/archive/%s.tar.gz' % version,
     packages=['phi',
-              'phi.app',
-              'phi.app._dash',
+              'phi.viz',
+              'phi.viz._dash',
               'phi.field',
               'phi.geom',
               'phi.math',
